+++ processing/init_raw_fluorescence_zarr.py
@@ -124,7 +124,7 @@ def create_output_zarr(output_path: str, num_pixels: int, num_cells: int, batch_
         output_path, 'acquisition_time_ms',
         shape=(num_pixels, SIZE_T),
         chunks=(pixel_chunk, batch_size),
-        dtype='uint32',
+        dtype='uint16',
     )
 
     # Static arrays: single chunk
